globalscholarshiphub/scraper/requirements.txt
scrapy>=2.11.0
beautifulsoup4>=4.12.0
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dateutil>=2.8.2
pydantic>=2.5.0
python-dotenv>=1.0.0
lxml>=4.9.3
fake-useragent>=1.4.0
selenium>=4.15.0
webdriver-manager>=4.0.1
